src/requirements.txt
annotated-types==0.7.0
anyio==4.4.0
blinker==1.8.2
certifi==2024.6.2
click==8.1.7
distro==1.9.0
feedgen==1.0.0
feedparser==6.0.11
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
greenlet==3.0.3
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
itsdangerous==2.2.0
Jinja2==3.1.4
joblib==1.4.2
lxml==5.2.2
MarkupSafe==2.1.5
numpy==1.26.4
openai==1.32.0
pandas==2.2.2
pydantic==2.7.3
pydantic_core==2.18.4
PyRSS2Gen==1.1
python-dateutil==2.9.0.post0
pytz==2024.1
PyYAML==6.0.1
scikit-learn==1.5.0
scipy==1.13.1
sgmllib3k==1.0.0
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.30
threadpoolctl==3.5.0
tqdm==4.66.4
typing_extensions==4.12.1
tzdata==2024.1
Werkzeug==3.0.3
